初步提交

+++ app/src/main/java/com/jack/jackzkrt/dialog/SettingEditTextDialog.java
@@ -84,7 +84,7 @@ public class SettingEditTextDialog extends AbsDialog {
                     return;
                 }
 
-                float number = Float.parseFloat(name);
+                int number = Integer.parseInt(name);
                 if(number>=mStart && number<=mStop) {
                     mCallBackExitText.exitName(number);
                     dismiss();
